backend.tsからstore周りの処理を移しました。 (#1450)
