+++ src/background.ts
@@ -14,22 +14,17 @@ import {
   net,
 } from "electron";
 import installExtension, { VUEJS_DEVTOOLS } from "electron-devtools-installer";
-import Store, { Schema } from "electron-store";
 import dotenv from "dotenv";
 
 import log from "electron-log";
 import dayjs from "dayjs";
 import windowStateKeeper from "electron-window-state";
-import zodToJsonSchema from "zod-to-json-schema";
 import { hasSupportedGpu } from "./electron/device";
 import {
   HotkeySetting,
   ThemeConf,
-  AcceptTermsStatus,
   EngineInfo,
-  ElectronStoreType,
   SystemError,
-  electronStoreSchema,
   defaultHotkeySettings,
   isMac,
   defaultToolbarButtonSetting,
@@ -52,6 +47,7 @@ import VvppManager, { isVvppFile } from "./background/vvppManager";
 import configMigration014 from "./background/configMigration014";
 import { failure, success } from "./type/result";
 import { ipcMainHandle, ipcMainSend } from "@/electron/ipc";
+import { getStoreWithError } from "@/background/electronStore";
 
 type SingleInstanceLockData = {
   filePath: string | undefined;
@@ -136,91 +132,7 @@ protocol.registerSchemesAsPrivileged([
 const firstUrl = process.env.VITE_DEV_SERVER_URL ?? "app://./index.html";
 
 // 設定ファイル
-const electronStoreJsonSchema = zodToJsonSchema(electronStoreSchema);
-if (!("properties" in electronStoreJsonSchema)) {
-  throw new Error("electronStoreJsonSchema must be object");
-}
-let store: Store<ElectronStoreType>;
-try {
-  store = new Store<ElectronStoreType>({
-    schema: electronStoreJsonSchema.properties as Schema<ElectronStoreType>,
-    migrations: {
-      ">=0.13": (store) => {
-        // acceptTems -> acceptTerms
-        const prevIdentifier = "acceptTems";
-        const prevValue = store.get(prevIdentifier, undefined) as
-          | AcceptTermsStatus
-          | undefined;
-        if (prevValue) {
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          store.delete(prevIdentifier as any);
-          store.set("acceptTerms", prevValue);
-        }
-      },
-      ">=0.14": (store) => {
-        // FIXME: できるならEngineManagerからEnginIDを取得したい
-        if (process.env.DEFAULT_ENGINE_INFOS == undefined)
-          throw new Error("DEFAULT_ENGINE_INFOS == undefined");
-        const engineId = EngineId(
-          JSON.parse(process.env.DEFAULT_ENGINE_INFOS)[0].uuid
-        );
-        if (engineId == undefined)
-          throw new Error("DEFAULT_ENGINE_INFOS[0].uuid == undefined");
-        const prevDefaultStyleIds = store.get("defaultStyleIds");
-        store.set(
-          "defaultStyleIds",
-          prevDefaultStyleIds.map((defaultStyle) => ({
-            engineId,
-            speakerUuid: defaultStyle.speakerUuid,
-            defaultStyleId: defaultStyle.defaultStyleId,
-          }))
-        );
-
-        const outputSamplingRate: number =
-          // @ts-expect-error 削除されたパラメータ。
-          store.get("savingSetting").outputSamplingRate;
-        store.set(`engineSettings.${engineId}`, {
-          useGpu: store.get("useGpu"),
-          outputSamplingRate:
-            outputSamplingRate === 24000 ? "engineDefault" : outputSamplingRate,
-        });
-        // @ts-expect-error 削除されたパラメータ。
-        store.delete("savingSetting.outputSamplingRate");
-        // @ts-expect-error 削除されたパラメータ。
-        store.delete("useGpu");
-      },
-    },
-  });
-} catch (e) {
-  log.error(e);
-  app.whenReady().then(() => {
-    dialog
-      .showMessageBox({
-        type: "error",
-        title: "設定ファイルの読み込みエラー",
-        message: `設定ファイルの読み込みに失敗しました。${app.getPath(
-          "userData"
-        )} にある config.json の名前を変えることで解決することがあります（ただし設定がすべてリセットされます）。設定ファイルがあるフォルダを開きますか？`,
-        buttons: ["いいえ", "はい"],
-        noLink: true,
-        cancelId: 0,
-      })
-      .then(async ({ response }) => {
-        if (response === 1) {
-          await shell.openPath(app.getPath("userData"));
-          // 直後にexitするとフォルダが開かないため
-          await new Promise((resolve) => {
-            setTimeout(resolve, 500);
-          });
-        }
-      })
-      .finally(() => {
-        app.exit(1);
-      });
-  });
-  throw e;
-}
-
+const store = getStoreWithError();
 // engine
 const vvppEngineDir = path.join(app.getPath("userData"), "vvpp-engines");
